Add files via upload
Mano o mario agora anda fluidamente :)

+++ DataStruct.hs
@@ -4,7 +4,7 @@ module DataStruct where
 data Mapa = Mapa (Posicao, Direcao) Posicao [[Bloco]]
 data Bloco = Escada | Plataforma | Alcapao | Vazio deriving (Show,Eq,Read)
 data Direcao = Norte | Sul | Leste | Oeste deriving (Show,Eq)
-type Posicao = (Float,Float)          --par ordenado que define (coluna,linha)
+type Posicao = (Float,Float)            --par ordenado que define (coluna,linha)
 data Personagem = Personagem{   
 velocidade :: Velocidade                --Um vetor de velocidade normal
 , tipo :: Entidade                      --Indica se é um inimigo (e qual) ou o jogador
@@ -17,7 +17,7 @@ velocidade :: Velocidade                --Um vetor de velocidade normal
 , pontos :: Int                         --Pontuaçao acumulada
 , aplicaDano :: (Bool, Double)          --Indica se tem o martelo armado e por quanto tempo ainda. Quando ativo, cria uma hitbox à frente do jogador (com as dimensoes do jogador) e qualquer inimigo que encoste sofre dano
 }
-type Velocidade = Double 
+type Velocidade = Float
 data Entidade = Jogador | MacacoMalvado | Fantasma deriving (Show,Eq,Read)
 data Colecionavel = Martelo | Moeda deriving (Show,Eq,Read)
